refactor: update extraId type to Long in IPK DTOs and add mappers
This commit changes the data type of the `extraId` field from `Int` to `Long` in `IpkReportRequest`, `IpkReportData`, `IpkBapRequest`, and `IpkBapReportData` DTOs.

Additionally, this commit introduces new mapper functions in `IpkDtoMapper.kt` for converting between domain models (`InspectionWithDetailsDomain`) and IPK (Instalasi Proteksi Kebakaran) DTOs for both BAP (Berita Acara Pemeriksaan) and Laporan (Report).

Key changes:
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/ipk/IpkReportDto.kt`:
    - Changed `extraId` type from `Int` to `Long` in `IpkReportRequest` and `IpkReportData`.
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/ipk/IpkBapDto.kt`:
    - Changed `extraId` type from `Int` to `Long` in `IpkBapRequest` and `IpkBapReportData`.
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/mapper/IpkDtoMapper.kt`:
    - Added `toIpkBapRequest()`: Converts `InspectionWithDetailsDomain` to `IpkBapRequest`.
    - Added `toInspectionWithDetailsDomain()` (for BAP): Converts `IpkBapReportData` to `InspectionWithDetailsDomain`.
    - Added `toIpkReportRequest()`: Converts `InspectionWithDetailsDomain` to `IpkReportRequest`.
    - Added `toInspectionWithDetailsDomain()` (for Report): Converts `IpkReportData` to `InspectionWithDetailsDomain`.
    - Includes mapping for general data, technical data, document checklists, visual inspections, functional tests, alarm systems, hydrant systems, and findings/recommendations.
    - Defines internal categories (`FireProtectionBAPCategory`, `FireProtectionCategory`) for consistent mapping logic.

+++ app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/ipk/IpkReportDto.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class IpkReportRequest(
     @SerializedName("examinationType") val examinationType: String,
-    @SerializedName("extraid") val extraId: Int,
+    @SerializedName("extraid") val extraId: Long,
     @SerializedName("createdAt") val createdAt: String,
     @SerializedName("inspectionType") val inspectionType: String,
     @SerializedName("equipmentType") val equipmentType: String,
@@ -37,7 +37,7 @@ data class IpkListReportResponseData(
 data class IpkReportData(
     @SerializedName("id") val id: String,
     @SerializedName("examinationType") val examinationType: String,
-    @SerializedName("extraid") val extraId: Int,
+    @SerializedName("extraid") val extraId: Long,
     @SerializedName("createdAt") val createdAt: String,
     @SerializedName("inspectionType") val inspectionType: String,
     @SerializedName("equipmentType") val equipmentType: String,
